% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCmisc.R
\name{comma.list}
\alias{comma.list}
\title{Print out comma separated list of values in X, truncating if many (good for error messages)}
\usage{
comma.list(X)
}
\arguments{
\item{X}{a vector to be displayed}
}
\value{
string with entries separated by commas, and if long, entries skipped
indicated by an ellipsis.
}
\description{
Often for nice presentation of error messages you wish to display a list of values.
This adds commas between entries and will truncate the list above a length of 50
items with an ellipsis. Very simple but convenient function.
}
\examples{
comma.list(1:100)
cat("The following entries were ignored: ", comma.list(c(1,7,10:14)), "\n")
}
