% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCmisc.R
\name{memory.summary}
\alias{memory.summary}
\title{Summary of RAM footprint for all R objects in the current session.
Not my function, but taken from an R-Help response by Elizabeth Purdom,
at Berkeley. Simply applies the function 'object.size' to the objects
in ls(). Also very similar to an example in the 'Help' for the 
utils::object.size() function.}
\usage{
memory.summary(unit = c("kb", "mb", "gb", "b"))
}
\arguments{
\item{unit}{default is to display "kb", but you can also choose
"b"=bytes, "mb"= megabyte, or "gb" = gigabytes. Only the first
letter is used, and is not case sensitive, so enter units how you
like.}
}
\value{
a list of object names with memory usage in bytes
}
\description{
Summary of RAM footprint for all R objects in the current session.
Not my function, but taken from an R-Help response by Elizabeth Purdom,
at Berkeley. Simply applies the function 'object.size' to the objects
in ls(). Also very similar to an example in the 'Help' for the 
utils::object.size() function.
}
\examples{
memory.summary() # shows memory used by all objects in the current session in kb
memory.summary("mb") # change units to megabytes
}

