% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NCmisc.R
\name{p.to.Z}
\alias{p.to.Z}
\title{Convert p-values to Z-scores}
\usage{
p.to.Z(p)
}
\arguments{
\item{p}{p-values (between 0 and 1), numeric, scalar, vector or matrix,
or other types coercible using as.numeric()}
}
\value{
Z scores with the same dimension as the input
}
\description{
Simple conversion of two-tailed p-values to Z-scores. Written
in a way that allows maximum precision for small p-values.
}
\examples{
p.to.Z(0.0001)
p.to.Z("5E-8")
p.to.Z(c(".05",".01",".005"))
p.to.Z(matrix(runif(16),nrow=4))
}
\author{
Nicholas Cooper \email{nick.cooper@cimr.cam.ac.uk}
}
\seealso{
\code{\link{Z.to.p}}
}

