% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/NCmisc.R
\name{pctile}
\alias{pctile}
\title{Find data thresholds corresponding to percentiles}
\usage{
pctile(dat, pc = 0.01)
}
\arguments{
\item{dat}{numeric vector of data}

\item{pc}{the percentile to seek, c(pc, 1-pc)}
}
\value{
returns the upper and lower threshold
}
\description{
Finds the top and bottom bounds corresponding to percentile
'pc' of the data 'dat'.
}
\examples{
pctile(rnorm(100000),.025)
pctile(sample(100),.9)
}
\author{
Nicholas Cooper \email{nick.cooper@cimr.cam.ac.uk}
}

