% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCmisc.R
\name{rmv.names}
\alias{rmv.names}
\title{Remove names from a named vector or list}
\usage{
rmv.names(X)
}
\arguments{
\item{X}{object for which you want to remove name}
}
\value{
the original object but without names
}
\description{
Convenience function, it's very easy to set names to NULL, but
this requires a dedicated line of code. Using this function
can make your code simpler.
}
\examples{
x <- c(boo=1, hiss=2)
rmv.names(x)
X <- list(testing=c(1,2,3), thankyou=TRUE)
rmv.names(X)
}
\author{
Nicholas Cooper \email{njcooper@gmx.co.uk}
}
