% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCmisc.R
\name{simple.date}
\alias{simple.date}
\title{Simple representation and retrieval of Date/Time}
\usage{
simple.date(sep = "_", long = FALSE, time = TRUE)
}
\arguments{
\item{sep}{character, separator to use for the date/time, eg, 
underscore or <space> " ".}

\item{long}{logical, whether to display a longer version of the
date and time, or just a simple version}

\item{time}{logical, whether to include the time, or just the date}
}
\value{
A string containing the date: MMMDD and optionally time HRam/pm.
Or if long=TRUE, a longer representation: DAY MM DD HH.MM.SS YYYY.
}
\description{
Retrieve a simple representation of date_time or just date, 
for generating day/time specific file names, etc.
}
\examples{
simple.date()
simple.date(" ",long=TRUE)
simple.date(time=FALSE)
}

